+++ handshake/handshake_message.go
@@ -39,7 +39,7 @@ func ParseHandshakeMessage(r utils.ReadStream) (Tag, map[Tag][]byte, error) {
 	var dataStart uint32
 	for indexPos := 0; indexPos < int(nPairs)*8; indexPos += 8 {
 		tag := Tag(binary.LittleEndian.Uint32(index[indexPos : indexPos+4]))
-		dataEnd := uint32(binary.LittleEndian.Uint32(index[indexPos+4 : indexPos+8]))
+		dataEnd := binary.LittleEndian.Uint32(index[indexPos+4 : indexPos+8])
 
 		dataLen := dataEnd - dataStart
 		if dataLen > protocol.CryptoParameterMaxLength {
